requirements.txt
Django~=4.1.2
djangorestframework~=3.14.0
jupyter~=1.0.0
pip~=21.0.1
attrs~=22.1.0
toml~=0.10.2
wheel~=0.36.2
response~=0.5.0
cryptography~=3.4.7
Pillow~=9.3.0
docutils~=0.17
Sphinx~=4.0.1
tornado~=6.1
Jinja2~=2.11.3
keyring~=23.9.3
pkginfo~=1.8.3
setuptools~=52.0.0
packaging~=21.3
pyparsing~=3.0.9
zipp~=3.4.1
Cython~=0.29.23
future~=0.18.2